backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langchain-core
pydantic
python-dotenv
